fix: box_iou_batch int-dtype overflow, narrow float32 precision claim
[resolve group] PR #2418 — items 1, 5, 9, 11

---

+++ src/supervision/detection/utils/iou_and_nms.py
@@ -226,8 +226,17 @@ def box_iou_batch(
         ```
     """
     overlap_metric = OverlapMetric.from_value(overlap_metric)
-    x_min_true, y_min_true, x_max_true, y_max_true = boxes_true.T
-    x_min_det, y_min_det, x_max_det, y_max_det = boxes_detection.T
+    # Upcast the corners to float64 right after unpacking so every subtraction
+    # and multiplication below runs in float64. This prevents integer-dtype
+    # overflow: an int32 box area such as 50000 * 50000 = 2.5e9 wraps to a
+    # negative value in int32 before any later cast could run, yielding wrong
+    # (often zero) IoU. Upcasting here also gives full float64 precision to
+    # float64/int64 callers. It does NOT recover precision already lost when a
+    # caller stores coordinates as float32 upstream, because that float32
+    # rounding happens before this function is ever called. The final matrix is
+    # cast back to float32 to preserve the public return-type contract.
+    x_min_true, y_min_true, x_max_true, y_max_true = boxes_true.T.astype(np.float64)
+    x_min_det, y_min_det, x_max_det, y_max_det = boxes_detection.T.astype(np.float64)
     count_true, count_det = boxes_true.shape[0], boxes_detection.shape[0]
 
     if count_true == 0 or count_det == 0:
@@ -235,10 +244,6 @@ def box_iou_batch(
             npt.NDArray[np.float32], np.empty((count_true, count_det), dtype=np.float32)
         )
 
-    # Accumulate in float64: float32 cannot represent integer coordinates above
-    # 2**24 exactly (e.g. GeoTIFF-scale pixel coordinates), which corrupts the
-    # min/max corners and the derived areas. The final matrix is cast back to
-    # float32 to preserve the public return-type contract.
     x_min_inter = np.empty((count_true, count_det), dtype=np.float64)
     x_max_inter = np.empty_like(x_min_inter)
     y_min_inter = np.empty_like(x_min_inter)
@@ -257,10 +262,8 @@ def box_iou_batch(
 
     area_inter = x_max_inter * y_max_inter  # inter_w * inter_h
 
-    area_true = ((x_max_true - x_min_true) * (y_max_true - y_min_true)).astype(
-        np.float64
-    )
-    area_det = ((x_max_det - x_min_det) * (y_max_det - y_min_det)).astype(np.float64)
+    area_true = (x_max_true - x_min_true) * (y_max_true - y_min_true)
+    area_det = (x_max_det - x_min_det) * (y_max_det - y_min_det)
 
     if overlap_metric == OverlapMetric.IOU:
         area_norm = area_true[:, None] + area_det[None, :] - area_inter
